requirements.txt
altgraph==0.17.3
fbs @ file:///C:/Users/S%C3%A9bastien/Documents/data_science/python/gui_qt_python/9-12-fbs_warmup/fbs_pro-1.1.4.tar.gz
future==0.18.2
pefile==2022.5.30
Pillow==9.2.0
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
PySide6==6.3.2
PySide6-Addons==6.3.2
PySide6-Essentials==6.3.2
pywin32-ctypes==0.2.0
shiboken6==6.3.2
